refactor: move parse_key_file to utils
