+++ crates/jstz_utils/src/lib.rs
@@ -1,8 +1,8 @@
-use jstz_crypto::{public_key::PublicKey, secret_key::SecretKey};
-use serde::Serialize;
 pub mod event_stream;
 pub mod filtered_log_stream;
+pub mod key_pair;
 pub mod tailed_file;
+pub use key_pair::KeyPair;
 
 pub async fn poll<'a, F, T>(
     max_attempts: u16,
@@ -22,19 +22,11 @@ where
     None
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize)]
-#[serde(into = "PublicKey")]
-pub struct KeyPair(pub PublicKey, pub SecretKey);
-
-impl From<KeyPair> for PublicKey {
-    fn from(value: KeyPair) -> Self {
-        value.0
-    }
-}
-
 // WARNING: Should only be used in tests!
 pub mod test_util {
-    use super::*;
+    use crate::key_pair::KeyPair;
+    use jstz_crypto::{public_key::PublicKey, secret_key::SecretKey};
+
     // Global tokio instance to prevent races among v2 runtime tests
     pub static TOKIO: std::sync::LazyLock<tokio::runtime::Runtime> =
         std::sync::LazyLock::new(|| {
